requirements.txt
Pillow
pyserial
pyinstaller